Rename entities
