+++ feature-libs/product-configurator/rulebased/components/config/default-configurator-message-config.ts
@@ -1,0 +1,9 @@
+import { ConfiguratorMessageConfig } from './configurator-message.config';
+
+export const DefaultConfiguratorMessageConfig: ConfiguratorMessageConfig = {
+  productConfigurator: {
+    updateConfigurationMessage: {
+      waitingTime: 1000,
+    },
+  },
+};
