--- feature-libs/product-configurator/rulebased/components/config/default-message-config.ts
@@ -1,9 +1,0 @@
-import { MessageConfig } from './message.config';
-
-export const DefaultMessageConfig: MessageConfig = {
-  productConfigurator: {
-    updateConfigurationMessage: {
-      waitingTime: 1000,
-    },
-  },
-};
