+++ feature-libs/product-configurator/rulebased/components/config/index.ts
@@ -1,3 +1,3 @@
+export { ConfiguratorMessageConfig as MessageConfig } from './configurator-message.config';
 export { ConfiguratorUISettingsConfig } from './configurator-ui-settings.config';
-export { MessageConfig } from './message.config';
 import './augmented-config.model';
