+++ feature-libs/product-configurator/rulebased/components/update-message/configurator-update-message.component.ts
@@ -3,7 +3,7 @@ import { ConfiguratorRouterExtractorService } from '@spartacus/product-configura
 import { Observable, of } from 'rxjs';
 import { delay, distinctUntilChanged, map, switchMap } from 'rxjs/operators';
 import { ConfiguratorCommonsService } from '../../core/facade/configurator-commons.service';
-import { MessageConfig } from '../config/message.config';
+import { ConfiguratorMessageConfig } from '../config/configurator-message.config';
 
 @Component({
   selector: 'cx-configurator-update-message',
@@ -42,6 +42,6 @@ export class ConfiguratorUpdateMessageComponent {
   constructor(
     protected configuratorCommonsService: ConfiguratorCommonsService,
     protected configRouterExtractorService: ConfiguratorRouterExtractorService,
-    protected config: MessageConfig
+    protected config: ConfiguratorMessageConfig
   ) {}
 }
